Add transaction.atomic() for write queryes in all services

+++ src/auction/controllers/lot/lotController.py
@@ -2,6 +2,7 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
 from auction.models.lot import Lot
+from rest_framework import status, serializers
 from auction.models.auction import Auction
 from auction.services.lot.lotService import LotService
 from django.contrib.auth.models import User
@@ -15,6 +16,9 @@ class LotController:
     @api_view(['POST'])
     def create(request):
         try:
+            if not request.user.is_authenticated or not request.user.is_active:
+                return Response({'message': 'Ошибка авторизации'}, status=status.HTTP_401_UNAUTHORIZED)
+            
             ownerId = request.user.id
             owner = get_object_or_404(User, id=ownerId)
             auctionId = request.data.get('auctionId')
@@ -25,42 +29,55 @@ class LotController:
             description = request.data.get('description')
             image = request.data.get('image')
 
-            response = LotController.lotService.create(owner, auction, startTime, endTime, title, description, image)
-
-            return Response(response, status=status.HTTP_201_CREATED)
-        except Lot.DoesNotExist:
-            raise ObjectDoesNotExist('Запрашиваемый лот не найден или не существует')
+            try:
+                response = LotController.lotService.create(owner, auction, startTime, endTime, title, description, image)
+                return Response(response, status=status.HTTP_201_CREATED)
+            except serializers.ValidationError as e:
+                return Response({'message': "Ошибка валидации", 'data': e.detail}, status=status.HTTP_400_BAD_REQUEST)
+            except Exception as e:
+                return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            
         except Exception as e:
-            return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            return Response({'message': str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     @api_view(['GET'])
     def getAll(request):
         try:
-            response = LotController.lotService.getAll()
-
-            return Response(response, status=status.HTTP_200_OK)
+            
+            try:
+                response = LotController.lotService.getAll()
+                return Response(response, status=status.HTTP_200_OK)
+            except Exception as e:
+                return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+    
         except Exception as e:
-            return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            return Response({'message': str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     @api_view(['GET'])
     def getById(request):
         try:
             lotId = request.data.get('lotId')
-
-            response = LotController.lotService.getById(lotId)
-
-            return Response(response, status=status.HTTP_200_OK)
-        except ObjectDoesNotExist as e:
+            
+            try:
+                response = LotController.lotService.getById(lotId)
+                return Response(response, status=status.HTTP_200_OK)
+            except ObjectDoesNotExist as e:
                 return Response({'message': str(e)}, status=status.HTTP_404_NOT_FOUND)
+            except Exception as e:
+                return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        
         except Exception as e:
-            return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            return Response({'message': str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     @api_view(['PATCH'])
     def update(request):
         try:
+            if not request.user.is_authenticated or not request.user.is_active:
+                return Response({'message': 'Ошибка авторизации'}, status=status.HTTP_401_UNAUTHORIZED)
+            
             lotId = request.data.get('lotId')
             startTime = request.data.get('startTime')
             endTime = request.data.get('endTime')
@@ -68,24 +85,35 @@ class LotController:
             description = request.data.get('description')
             image = request.data.get('image')
 
-            response = LotController.lotService.update(lotId, startTime, endTime, title, description, image)
-
-            return Response(response, status=status.HTTP_200_OK)
-        except ObjectDoesNotExist as e:
+            try:
+                response = LotController.lotService.update(lotId, startTime, endTime, title, description, image)
+                return Response(response, status=status.HTTP_200_OK)
+            except ObjectDoesNotExist as e:
                 return Response({'message': str(e)}, status=status.HTTP_404_NOT_FOUND)
+            except serializers.ValidationError as e:
+                return Response({'message': "Ошибка валидации", 'data': e.detail}, status=status.HTTP_400_BAD_REQUEST)
+            except Exception as e:
+                return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
         except Exception as e:
-            return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            return Response({'message': str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     @api_view(['DELETE'])
     def delete(request):
         try:
+            if not request.user.is_authenticated or not request.user.is_active:
+                return Response({'message': 'Ошибка авторизации'}, status=status.HTTP_401_UNAUTHORIZED)
+            
             lotId = request.data.get('lotId')
 
-            response = LotController.lotService.delete(lotId)
-
-            return Response(response, status=status.HTTP_200_OK)
-        except ObjectDoesNotExist as e:
+            try:
+                response = LotController.lotService.delete(lotId)
+                return Response(response, status=status.HTTP_200_OK)
+            except ObjectDoesNotExist as e:
                 return Response({'message': str(e)}, status=status.HTTP_404_NOT_FOUND)
+            except Exception as e:
+                return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
         except Exception as e:
-            return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            return Response({'message': str(e)}, status=status.HTTP_400_BAD_REQUEST)
